Share data volume formatting helper for data-driven graphs
